requirements.txt
Flask~=3.0.2
cryptography~=42.0.5
pycryptodome
Pillow~=10.0.0
boto3~=1.28.38
requests~=2.31.0
Gunicorn